refactor: split `deploy` into smaller functions
The smaller functions are easier to test and don't require as much mocking setup.
