+++ src/tests/mod.rs
@@ -1,6 +1,7 @@
 mod clean;
-mod deploy;
+mod create_infra;
 mod init;
+mod provision_genesis_node;
 mod setup;
 
 const RPC_CLIENT_BIN_NAME: &str = "safenode_rpc_client";
